[Cmake] Fix error that I made because I am stoopid
The linking of extern library was wrong (copy paste isn't enough and it
needed modification as a name wasn't the right one)
Applying as well the patch from another commit that has been force
pushed to the repository.
This commit will also be force pushed in order to reapply the repo
organization that had been made previously.
Force pushing commit is not something to use all the time so avoid using
it.
Adding as well a .clang-format file