[Movements] Fix Movements errors
Copying an item was a bad idea as it lead to a lot of problems and
mostly if it kept the same object then modifying one would also
modify the other.
The loop to move the snake was duplicated so it is a single loop and
instead of saving the compute time of one more lap in the for loop we do
now need to take this lap to set the coordinates of the newly added
body part of the snake.
Update asset of the apple to the right one instead of using test asset.
Delete unused test asset.