[Pacman] The game is now working
THe pacman game is working, the pacgums can be eaten and the score
increases, the teleportation is also working, the pacman moves and
rotates.
Some names have been reviewed in snake and some variables aren't
initialized with value that don't make sense
The Pacman map has been updated to correspond to what the game is like.
Add some more colors for ncurses.
Add some keywords to force people to use the return values of getters
and to prevent wrong types in constructors.
Add github actions workflows