[Snake] add timed game with std::chrono
The Snake is timed so that it moves 4 squares per seconds.
The documentation is prepared and a first pdf has been generated.
Remove usleep in core because std::chrono is now used in every game.
Use braced initialization on SFML variables where it wasn't used.
ignoring files generated with doxygen used to then generate the pdf.